+++ rpio/pyLauncher/pyLauncher.py
@@ -29,3 +29,17 @@ def launch_main(mainFile='main.py'):
     """
     command = ["python", mainFile]
     call(command)
+
+def launch_docker_compose(path='/'):
+    """Function to launch one or more python software component using a docker compose file
+
+            :param [path]: [path to the docker compose file for the given platform], defaults to [/]
+            :type [path]: [String](, optional)
+            ...
+            :return: [Functions returns nothing]
+            :rtype: [None]
+        """
+    cdCommand = ["cd", path]
+    call(cdCommand)
+    dockerCommand = ["docker compose up --build -d"]    #-d runs in detached mode
+    call(dockerCommand)